+++ eslint.config.mjs
@@ -79,6 +79,9 @@ export default tsEslint.config(
       'node-import/prefer-node-protocol': 'error',
       // https://typescript-eslint.io/rules/method-signature-style/
       '@typescript-eslint/method-signature-style': ['error', 'property'],
+      // https://typescript-eslint.io/rules/no-unnecessary-type-assertion/
+      // Legacy use - This rule is disabled because it conflicts with the 'consistent-type-assertions' rule.
+      '@typescript-eslint/no-unnecessary-type-assertion': 'off',
       // JSDoc overrides
       'jsdoc/tag-lines': [
         'error',
@@ -150,6 +153,16 @@ export default tsEslint.config(
       'jsdoc/require-param-type': 'off', // override of 'plugin:jsdoc/recommended'
       'jsdoc/require-returns-type': 'off', // override of 'plugin:jsdoc/recommended'
       'jsdoc/require-throws-type': 'off', // override of 'plugin:jsdoc/recommended'
+    },
+    settings: {
+      jsdoc: {
+        tagNamePreference: {
+          // https://typedoc.org/documents/Tags._template.html
+          // TypeDoc recognizes the @template tag as an alias of @typeParam for compatibility with JavaScript projects
+          // using TypeScript via doc comments. For TypeScript projects, the TSDoc standard @typeParam tag should be preferred.
+          template: 'typeParam',
+        },
+      },
     },
   },
   {